fix: resize clock icon to be same size as other icons

See merge request e/os/BlissLauncher!116

+++ app/src/main/java/foundation/e/blisslauncher/core/customviews/DialOverlay.java
@@ -20,7 +20,7 @@ public interface DialOverlay {
      * @param h        the height of the canvas
      * @param calendar the desired date/time
      */
-    void onDraw(Canvas canvas, int cX, int cY, int w, int h, Calendar calendar,
+    void onDraw(Canvas canvas, float cX, float cY, int w, int h, Calendar calendar,
             boolean sizeChanged);
 
 }
